Move out utility type families
