+++ Plutarch/Internal/PlutusType.hs
@@ -32,6 +32,7 @@ import Plutarch.DataRepr.Internal.HList (IndexList)
 import Plutarch.Internal (PType, S, Term, pforce, plam', punsafeCoerce, (:-->))
 import qualified Plutarch.Internal as PI
 import Plutarch.Internal.PLam ((#))
+import Plutarch.Internal.TypeFamily (ToPType, ToPType2)
 
 {- |
 
@@ -331,18 +332,3 @@ type ScottFn' :: [PType] -> PType -> PType
 type family ScottFn' xs b where
   ScottFn' '[] b = b
   ScottFn' (x ': xs) b = x :--> ScottFn' xs b
-
--- | Convert a list of `Term s a` to a list of `a`.
-type ToPType :: [Type] -> [PType]
-type family ToPType as where
-  ToPType '[] = '[]
-  ToPType (a ': as) = UnTerm a ': ToPType as
-
-type ToPType2 :: [[Type]] -> [[PType]]
-type family ToPType2 as where
-  ToPType2 '[] = '[]
-  ToPType2 (a ': as) = ToPType a ': ToPType2 as
-
-type UnTerm :: Type -> PType
-type family UnTerm x where
-  UnTerm (Term s a) = a
